SpringBootガイドのHandling Form Submissionを実施。 inputで受け取った値を画面上に表示するようにした。
ファイルが多くなってきたため、整理しました。
