+++ src/main/java/com/example/demo/BuildingRestfulWebService/BuildingRestfulWebService.java
@@ -1,0 +1,19 @@
+// package com.example.demo.BuildingRestfulWebService;
+
+// public class BuildingRestfulWebService {
+//   private final long id;
+//   private final String content;
+
+//   public BuildingRestfulWebService(long id, String content) {
+//     this.id = id;
+//     this.content = content;
+//   }
+
+//   public long getId() {
+//     return id;
+//   }
+
+//   public String getContent() {
+//     return content;
+//   }
+// }
