+++ src/main/java/com/example/demo/ConsumingRESTfulWebService/Value.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.ConsumingRESTfulWebService;
 
 public class Value {
