+++ src/main/java/com/example/demo/DemoApplication.java
@@ -3,10 +3,13 @@ package com.example.demo;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.context.annotation.Bean;
+	// SpringBootガイドのAccessing Data with JPAで実装。今回は不要な為コメントアウト。
+// import com.example.demo.entity.Customer;
+// import com.example.demo.repository.CustomerRepository;
+// import org.slf4j.Logger;
+// import org.slf4j.LoggerFactory;
+// import org.springframework.boot.CommandLineRunner;
+// import org.springframework.context.annotation.Bean;
 
 // SpringBootガイドのConsuming a RESTful Web Serviceで実装。今回は不要な為コメントアウト。
 // import org.springframework.boot.web.client.RestTemplateBuilder;
@@ -14,6 +17,10 @@ import org.springframework.context.annotation.Bean;
 
 @SpringBootApplication
 public class DemoApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DemoApplication.class, args);
+	}
 
 	// ============================================================================================================
 	// SpringBootガイドのConsuming a RESTful Web Serviceで実装。今回は不要な為コメントアウト。
@@ -33,52 +40,49 @@ public class DemoApplication {
 	// 	};
 	// }
 
+	// ============================================================================================================
+	// SpringBootガイドのAccessing Data with JPAで実装。今回は不要な為コメントアウト。
+	// ============================================================================================================
+	// private static final Logger log = LoggerFactory.getLogger(DemoApplication.class);
 
-	// SpringBootガイドのAccessing Data with JPAで実装。
-	private static final Logger log = LoggerFactory.getLogger(DemoApplication.class);
+	// @Bean
+  // public CommandLineRunner demo(CustomerRepository repository) {
+  //   return (args) -> {
+  //     // save a few customers
+  //     repository.save(new Customer("Jack", "Bauer"));
+  //     repository.save(new Customer("Chloe", "O'Brian"));
+  //     repository.save(new Customer("Kim", "Bauer"));
+  //     repository.save(new Customer("David", "Palmer"));
+  //     repository.save(new Customer("Michelle", "Dessler"));
 
-	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
-	}
+  //     // fetch all customers
+  //     log.info("findAll()で見つかった顧客:");
+  //     log.info("-------------------------------");
+  //     for (Customer customer : repository.findAll()) {
+  //       log.info(customer.toString());
+	// 		}
+	// 		log.info("-------------------------------");
+  //     log.info("");
 
-	@Bean
-  public CommandLineRunner demo(CustomerRepository repository) {
-    return (args) -> {
-      // save a few customers
-      repository.save(new Customer("Jack", "Bauer"));
-      repository.save(new Customer("Chloe", "O'Brian"));
-      repository.save(new Customer("Kim", "Bauer"));
-      repository.save(new Customer("David", "Palmer"));
-      repository.save(new Customer("Michelle", "Dessler"));
+  //     // fetch an individual customer by ID
+  //     Customer customer = repository.findById(1L);
+  //     log.info("findById(1L)で見つかった顧客:");
+  //     log.info("--------------------------------");
+	// 		log.info(customer.toString());
+	// 		log.info("-------------------------------");
+  //     log.info("");
 
-      // fetch all customers
-      log.info("findAll()で見つかった顧客:");
-      log.info("-------------------------------");
-      for (Customer customer : repository.findAll()) {
-        log.info(customer.toString());
-			}
-			log.info("-------------------------------");
-      log.info("");
-
-      // fetch an individual customer by ID
-      Customer customer = repository.findById(1L);
-      log.info("findById(1L)で見つかった顧客:");
-      log.info("--------------------------------");
-			log.info(customer.toString());
-			log.info("-------------------------------");
-      log.info("");
-
-      // fetch customers by last name
-      log.info("findByLastName('Bauer')で見つかった顧客:");
-      log.info("--------------------------------------------");
-      repository.findByLastName("Bauer").forEach(bauer -> {
-        log.info(bauer.toString());
-      });
-      // for (Customer bauer : repository.findByLastName("Bauer")) {
-      //  log.info(bauer.toString());
-      // }
-      log.info("");
-    };
-  }
+  //     // fetch customers by last name
+  //     log.info("findByLastName('Bauer')で見つかった顧客:");
+  //     log.info("--------------------------------------------");
+  //     repository.findByLastName("Bauer").forEach(bauer -> {
+  //       log.info(bauer.toString());
+  //     });
+  //     // for (Customer bauer : repository.findByLastName("Bauer")) {
+  //     //  log.info(bauer.toString());
+  //     // }
+  //     log.info("");
+  //   };
+  // }
 
 }
