--- src/main/java/com/example/demo/Greeting.java
@@ -1,19 +1,0 @@
-package com.example.demo;
-
-public class Greeting {
-  private final long id;
-  private final String content;
-
-  public Greeting(long id, String content) {
-    this.id = id;
-    this.content = content;
-  }
-
-  public long getId() {
-    return id;
-  }
-
-  public String getContent() {
-    return content;
-  }
-}
