+++ src/main/java/com/example/demo/HandlingFormSubmission/Greeting.java
@@ -1,0 +1,22 @@
+package com.example.demo.HandlingFormSubmission;
+
+public class Greeting {
+  private long id;
+  private String content;
+
+  public long getId() {
+    return id;
+  }
+
+  public void setId(long id) {
+    this.id = id;
+  }
+
+  public String getContent() {
+    return content;
+  }
+
+  public void setContent(String content) {
+    this.content = content;
+  }
+}
