+++ src/main/java/com/example/demo/controller/HandlingFormSubmission/HandlingFormSubmissionController.java
@@ -1,0 +1,26 @@
+package com.example.demo.controller.HandlingFormSubmission;
+
+import com.example.demo.HandlingFormSubmission.Greeting;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
+
+
+@Controller
+public class HandlingFormSubmissionController {
+
+  @GetMapping("/greeting")
+  public String greetingForm(Model model) {
+    model.addAttribute("greeting", new Greeting());
+    return "greeting";
+  }
+
+  @PostMapping("/greeting")
+  public String greetingSubmit(@ModelAttribute Greeting greeting) {
+    return "result";
+  }
+
+}
