+++ src/main/java/com/example/demo/repository/CustomerRepository.java
@@ -1,6 +1,9 @@
-package com.example.demo;
+package com.example.demo.repository;
 
 import java.util.List;
+
+import com.example.demo.entity.Customer;
+
 import org.springframework.data.repository.CrudRepository;
 
 public interface CustomerRepository extends CrudRepository<Customer, Long> {
